requirements.txt
# AWS → GCP Migration Project Dependencies
# Install: pip install -r requirements.txt

# AWS
boto3>=1.34.0

# GCP
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.14.0
google-cloud-bigquery-storage>=2.24.0

# Spark (local)
pyspark>=3.5.0

# Data generation
faker>=22.0.0
pyarrow>=15.0.0
pandas>=2.2.0

# Validation
jsonschema>=4.21.0

# Environment
python-dotenv>=1.0.0

# Airflow (local - install separately)
# pip install apache-airflow==2.8.1 apache-airflow-providers-google
